# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My API",
        "version": "1.0.0"
    },
    "paths": {
        "/api/app-version/latest": {
            "post": {
                "tags": [
                    "App Versions"
                ],
                "summary": "Flutter uygulamasına en son sürümü döndürür",
                "operationId": "9b3349ffbcdc1e1b5257766e2c2b5f73",
                "requestBody": {
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "platform": {
                                        "description": "android veya ios",
                                        "type": "string",
                                        "example": "android"
                                    },
                                    "version": {
                                        "description": "Mevcut kullanıcı sürümü",
                                        "type": "string",
                                        "example": "1.0.0"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "En son sürüm veya zaten en güncel sürüm",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "You are using the latest version"
                                        },
                                        "version": {
                                            "type": "string",
                                            "example": "2.0.0"
                                        },
                                        "url": {
                                            "type": "string",
                                            "example": "https://example.com/app.apk"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Sürüm bulunamadı",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "No version found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}